created List component
