+++ client/src/components/AddVideo.js
@@ -7,6 +7,7 @@ function AddVideo() {
   const [addVideo, setAddVideo] = React.useState("");
 
   console.log(addVideo);
+
   function handleAddVideo() {
     console.log("button clicked");
     const form = Form;
